+++ onlyabue/components/CitasProgramadas.jsx
@@ -1,0 +1,15 @@
+import { View, Text, } from "native-base";
+import {  StyleSheet} from "react-native";
+import { useState } from "react";
+export function CitaCard(){
+    const [NombreMed, setNombreMed] =useState('')
+    
+    return(
+        <View>
+            <Text>
+
+            </Text>
+            
+        </View>
+    );
+}
